third_party/blink/web_tests/external/wpt/css/css-values/if-style-query-expected.txt

This is a testharness.js-based test.
Found 13 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Values and Units Test: CSS inline if() function 10
  assert_equals: "if(style(--x: initial): true_value; else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 11
  assert_equals: "if(style(--x: initial): true_value; else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 13
  assert_equals: "if(style(--inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 14
  assert_equals: "if(style(--inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 21
  assert_equals: "if(style(--x: initial): true_value; else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 22
  assert_equals: "if(style(--x: initial): true_value; else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 24
  assert_equals: "if(style(--inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 25
  assert_equals: "if(style(--inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 67
  assert_equals: "if(style(--length: initial): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 68
  assert_equals: "if(style(--length: initial): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 70
  assert_equals: "if(style(--length: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 71
  assert_equals: "if(style(--length-inherited: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
[FAIL] CSS Values and Units Test: CSS inline if() function 72
  assert_equals: "if(style(--length: inherit): true_value;\n          else: false_value)" should be substituted to "true_value". expected "true_value" but got "false_value"
Harness: the test ran to completion.
